bpf: Fix build warning in conntrack test
A long, long time ago in a galaxy not so far from here...

  $ make -C test
    GINKG testbuild
  Compiling test...
      compiled test.test
    CC    test/bpf/elf-demo.o
    CC    test/bpf/unit-test
  In file included from unit-test.c:20:
  ../../bpf/lib/conntrack_test.h:16:13: warning: cast to smaller integer type
  'int' from 'void *' [-Wvoid-pointer-to-int-cast]
                  int idx = (int)tuple;
                            ^~~~~~~~~~
  1 warning generated.

+++ bpf/lib/conntrack_test.h
@@ -13,7 +13,7 @@ static struct ct_entry __ipv4_map[] = {
 static void *__map_lookup_elem(void *map, void *tuple)
 {
 	if (map == __ipv4_map) {
-		int idx = (int)tuple;
+		__u64 idx = (__u64)tuple;
 		if (idx == __TUPLE_EXIST)
 			return &__ipv4_map[idx];
 	}
